Fastclick fix: mcasimir/mobile-angular-ui#230
Investigated this, and I think it has never worked correctly. The
recommended way of attaching Fastclick isn't working with AngularJS.

It is being attached to the DOMContentloaded event, which in a SPA fires
only once and after that angular is starting:
> Angular initializes automatically upon DOMContentLoaded event or when
the angular.js script is evaluated if at that time document.readyState
is set to 'complete'. At this point Angular looks for the ng-app
directive which designates your application root.